rename(MyPage): 마이페이지 관련 파일 MyPage로 이동
